TEST-UNIFY-2: consolidate tests/helpers → tests/lib + import codemod (Phase 2)
The reusable test layer now lives in ONE place. Moved all helpers (auth, api,
seedGraph, dbHealing, testEnv, mobileAudit, zorder, vlm, mock-oauth-server, README)
into tests/lib/ alongside the harness modules, and balance_metrics.py into
tests/lib/metrics/. Codemod across all 39 keeper specs: '../helpers/' → '../lib/'
(uniform — every spec is depth-1 under tests/); fixed the balance_metrics.py path
string in graph-balance.spec.ts and two stale doc-comment paths.

Verified (the plan's #1 risk — a missed import silently zeroes a project at
collection): playwright --list per project all resolve with ZERO collection errors
(default 352, perf 3, geometry 18, showcase 5, vlm 1, matrix 109); THE GATE 5/5.
