+++ tests/diagnostics/core-interactions.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, Page } from '@playwright/test';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * CORE INTERACTION MATRIX — the "basic user checks" every build must pass.
