+++ tests/diagnostics/hierarchy-navigation.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect, Page } from '@playwright/test';
 import * as fs from 'fs';
 import * as path from 'path';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * Altium-style "graphs of graphs" navigation: from the System Overview, a
