+++ tests/diagnostics/large-graph-profile.spec.ts
@@ -1,7 +1,7 @@
 import { test, expect, Page } from '@playwright/test';
 import * as fs from 'fs';
 import * as path from 'path';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * BASELINE profiler for the Compute Core (1000-node) example graph the user
