+++ tests/diagnostics/minimap-zoom.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, Page } from '@playwright/test';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * Minimap wheel/pinch zoom: a wheel over the minimap zooms the MAIN view
