+++ tests/diagnostics/node-expand-legibility.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, Page } from '@playwright/test';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * PR-3 (expand-in-place) + PR-4 (zoom-decoupled legibility floor).
