+++ tests/e2e/matrix.setup.ts
@@ -1,5 +1,5 @@
 import { test as setup } from '@playwright/test';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 // One real login for the whole matrix run; every feature-matrix test reuses this
 // storage state so the ~100-cell resolution×feature sweep doesn't re-login each
