+++ tests/e2e/oauth-provider-config.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from '@playwright/test';
-import { login, TEST_USERS, getBaseURL } from '../helpers/auth';
+import { login, TEST_USERS, getBaseURL } from '../lib/auth';
 
 test.describe('OAuth Provider Configuration (Admin Panel)', () => {
   test.beforeEach(async ({ page }) => {
