+++ tests/lib/auth.ts
