+++ tests/perf/graph-perf.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect, Page } from '@playwright/test';
-import { login, TEST_USERS } from '../helpers/auth';
+import { login, TEST_USERS } from '../lib/auth';
 
 /**
  * Runtime performance budgets (ADAPT-8). Numbers come from the in-app
